+++ Grbl844Mega2560/print.h
@@ -20,7 +20,7 @@
 */
 
 /* This code was initially inspired by the wiring_serial module by David A. Mellis which
-   used to be a part of the Arduino project. */ 
+   used to be a part of the Arduino project. */
 
 #ifndef print_h
 #define print_h
@@ -33,6 +33,6 @@ void printInteger(long n);
 
 void print_uint8_base2(uint8_t n);
 
-void printFloat(double n);
+void printFloat(float n);
 
 #endif
